Move config stuff to config.go
